+++ src/VaultSharp/V1/AuthMethods/AliCloud/IAliCloudAuthMethod.cs
@@ -1,0 +1,9 @@
+﻿namespace VaultSharp.V1.AuthMethods.AliCloud
+{
+    /// <summary>
+    /// Non login operations.
+    /// </summary>
+    public interface IAliCloudAuthMethod
+    {
+    }
+}